# Add `sort!` and `minmax` to `Performance/CompareWithBlock`