--- a/changelog/change_compare_with_block_methods.md
+++ b/changelog/change_compare_with_block_methods.md
@@ -1,0 +1,1 @@
+* [#357](https://github.com/rubocop/rubocop-performance/pull/357): Add `sort!` and `minmax` to `Performance/CompareWithBlock`. ([@vlad-pisanov][])