`Performance/CompareWithBlock` currently only looks at `sort`, `min` and `max` usage.

This PR adds support for `sort!` and `minmax` methods which behave the same way.

```ruby
# bad
arr.minmax { |a, b| a.foo <=> b.foo }
arr.sort!  { |a, b| a.foo <=> b.foo }

# good
arr.minmax_by(&:foo)
arr.sort_by!(&:foo)
```

```
Comparison:
               minmax_by:  1513625.0 i/s
               minmax:     1178249.9 i/s - 1.28x  slower
```

```
Comparison:
               sort_by!:    21368.8 i/s
               sort!:       13662.7 i/s - 1.56x  slower
```

-----------------

Before submitting the PR make sure the following are checked:

* [x] The PR relates to *only* one subject with a clear title and description in grammatically correct, complete sentences.
* [x] Wrote [good commit messages][1].
* [ ] Commit message starts with `[Fix #issue-number]` (if the related issue exists).
* [x] Feature branch is up-to-date with `master` (if not - rebase it).
* [x] Squashed related commits together.
* [x] Added tests.
* [x] Ran `bundle exec rake default`. It executes all tests and runs RuboCop on its own code.
* [x] Added an entry (file) to the [changelog folder](https://github.com/rubocop/rubocop-performance/blob/master/changelog/) named `{change_type}_{change_description}.md` if the new code introduces user-observable changes. See [changelog entry format](https://github.com/rubocop/rubocop/blob/master/CONTRIBUTING.md#changelog-entry-format) for details.

[1]: https://chris.beams.io/posts/git-commit/